Disable value animation at start up % update FuelLevel path.
Also:
 - Remove some warnings.
 - Improvements of hvac temp buttons.

Original from: Dominik Wawrzonek <dwawrzonek@ics.com>

Bug-AGL: SPEC-4971

+++ lib/presentation/screens/dashboard/widgets/circle_indicator.dart
@@ -285,7 +285,7 @@ class FuelProgressIndicatorState extends ConsumerState<FuelProgressIndicator>
                 width: 220,
                 child: CustomPaint(
                   foregroundPainter: CirclePainter(
-                      value: fuelLevel,
+                      value: fuelLevel.toDouble(),
                       maxValue: maxFuelLevel,
                       isFuel: true,
                       isRPM: false),
